refactor: move dummy data in one folder and initialized a DataTable component
