+++ frontend/src/components/DataTable/index.js
@@ -1,0 +1,11 @@
+import './index.css';
+
+const DataTable = ({ data }) => {
+  return (
+    <div>
+      DataTable
+    </div>
+  );
+}
+
+export default DataTable;
